+++ app/components/Popup/TransactionState/ConfirmTransactionState.js
@@ -2,7 +2,7 @@ import React from 'react'
 
 import mobileImage from 'assets/images/mobile.svg'
 import styles from 'assets/css/global.css'
-import { CONFIRM_TRANSACTION_STATE } from '../../../../../../../config/messages'
+import { CONFIRM_TRANSACTION_STATE } from '../../../../config/messages'
 
 const ConfirmTransactionState = () => (
   <div className={styles.transactionState}>
